Put Table related files into a separate folder
